BCM270X_DT: Move vc4 node contents to bcm2708_common.dtsi.
This should clarify what's going on with the overlay: The hardware is
always present, we're just enabling the DT node so that the vc4 driver
probes.

The interrupts are left in the overlay, because the firmware doesn't
check node status before masking out the vc4 interrupts.

By having the nodes in the common file, we'll be able to correctly
connect the HDMI HPD GPIO so that we can detect whether an HDMI
monitor is connected.

Signed-off-by: Eric Anholt <eric@anholt.net>