kernek/fork.c: allocate idle task for a CPU always on its local node
BugLink: http://bugs.launchpad.net/bugs/1585850

Linux preallocates the task structs of the idle tasks for all possible
CPUs.  This currently means they all end up on node 0.  This also implies
that the cache line of MWAIT, which is around the flags field in the task
struct, are all located in node 0.

We see a noticeable performance improvement on Knights Landing CPUs when
the cache lines used for MWAIT are located in the local nodes of the CPUs
using them.  I would expect this to give a (likely slight) improvement on
other systems too.

The patch implements placing the idle task in the node of its CPUs, by
passing the right target node to copy_process()

[akpm@linux-foundation.org: use NUMA_NO_NODE, not a bare -1]
Link: http://lkml.kernel.org/r/1463492694-15833-1-git-send-email-andi@firstfloor.org
Signed-off-by: Andi Kleen <ak@linux.intel.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>

(cherry picked from linux-next commit b97773886a6864fcac4039f8a12ed5177c2218ba)
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>